refactor: inline old 'newNote' code
