+++ libanki/src/test/java/com/ichi2/anki/libanki/CollectionTest.kt
@@ -20,6 +20,7 @@ import anki.notetypes.StockNotetype
 import com.ichi2.anki.libanki.testutils.InMemoryAnkiTest
 import com.ichi2.anki.libanki.testutils.ext.addNote
 import com.ichi2.anki.libanki.testutils.ext.createBasicNoteType
+import com.ichi2.anki.libanki.testutils.ext.newNote
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers
 import org.hamcrest.Matchers.equalTo
