+++ libanki/testutils/src/main/java/com/ichi2/anki/libanki/testutils/ext/Collection.kt
@@ -61,3 +61,11 @@ fun Collection.createBasicTypingNoteType(name: String): NotetypeJson {
     notetypes.save(noteType)
     return noteType
 }
+
+/**
+ * Return a new note with the model derived from the deck or the configuration
+ * @param forDeck When true it uses the model specified in the deck (mid), otherwise it uses the model specified in
+ * the configuration (curModel)
+ * @return The new note
+ */
+fun Collection.newNote(forDeck: Boolean = true): Note = newNote(notetypes.current(forDeck))
